# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Relaxation Hub API
  description: API for Relaxation Hub - On-demand massage and relaxation services.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development
  - url: https://yourdomain.com/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [client, therapist, admin]
    Service:
      type: object
      properties:
        service_id:
          type: string
        name:
          type: string
        description:
          type: string
        base_price:
          type: number
        duration_minutes:
          type: integer
        category:
          type: string
    Address:
      type: object
      properties:
        address_id:
          type: string
        label:
          type: string
        street_address:
          type: string
        barangay:
          type: string
        city:
          type: string
        province:
          type: string
        postal_code:
          type: string
        landmark:
          type: string
        is_default:
          type: boolean
...
security:
  - BearerAuth: []
paths:
  /register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                provider:
                  type: string
                  enum: [email]
                provider_key:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum: [client, therapist]
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /login:
    post:
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [email]
                provider_key:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
  /oauth/{provider}:
    post:
      summary: Initiate OAuth login
      security: []
      parameters:
        - in: path
          name: provider
          schema:
            type: string
            enum: [google, apple]
          required: true
      responses:
        "302":
          description: Redirect to provider
  /oauth/callback:
    get:
      summary: OAuth callback
      security: []
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
        - in: query
          name: state
          schema:
            type: string
      responses:
        "200":
          description: OAuth login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
...
  /services:
    get:
      summary: List available services
      security: []
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
    post:
      summary: Create a new service (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Service"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /addresses:
    get:
      summary: List user addresses
      responses:
        "200":
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      summary: Add a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                street_address:
                  type: string
                barangay:
                  type: string
                city:
                  type: string
                province:
                  type: string
                postal_code:
                  type: string
                landmark:
                  type: string
                is_default:
                  type: boolean
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{id}:
    get:
      summary: Get address details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Address details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    patch:
      summary: Update address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                street_address:
                  type: string
                barangay:
                  type: string
                city:
                  type: string
                province:
                  type: string
                postal_code:
                  type: string
                landmark:
                  type: string
                is_default:
                  type: boolean
      responses:
        "200":
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      summary: Delete address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Address deleted
  /addresses/{id}/default:
    post:
      summary: Set address as default
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Address set as default
  /bookings:
    get:
      summary: List bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
    post:
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_id:
                  type: string
                address_id:
                  type: string
                scheduled_at:
                  type: string
                  format: date-time
                special_requests:
                  type: string
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      summary: Get booking details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      summary: Update booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                promotion_id:
                  type: string
                status:
                  type: string
                cancellation_reason:
                  type: string
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    post:
      summary: Update booking status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [in_progress, completed, refunded]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /payments:
    post:
      summary: Make a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                booking_id:
                  type: string
                payment_method:
                  type: string
                amount:
                  type: number
                payment_reference:
                  type: string
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /promotions:
    get:
      summary: List active promotions
      responses:
        "200":
          description: List of promotions
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Promotion"
    post:
      summary: Create a promotion (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                description:
                  type: string
                discount_percentage:
                  type: number
                valid_from:
                  type: string
                  format: date-time
                valid_until:
                  type: string
                  format: date-time
                max_uses:
                  type: integer
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
  /promotions/code:
    get:
      summary: Get promotion by code
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promotion details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
...
  /notifications:
    post:
      summary: Create a notification (Admin/System)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                title:
                  type: string
                message:
                  type: string
                type:
                  type: string
      responses:
        "201":
          description: Notification created
    get:
      summary: List notifications
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
...
  /messages/conversation:
    post:
      summary: Create conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participant_ids:
                  type: array
                  items:
                    type: string
                booking_id:
                  type: string
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /messages/send:
    post:
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: string
                message:
                  type: string
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /reviews:
    post:
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                therapist_id:
                  type: string
                booking_id:
                  type: string
                rating:
                  type: integer
                comment:
                  type: string
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /referrals:
    get:
      summary: List referrals
      responses:
        "200":
          description: List of referrals
          content:
            application/json:
              schema:
                type: object
                properties:
                  referrals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Referral"
    post:
      summary: Create a referral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                referred_user_email:
                  type: string
      responses:
        "201":
          description: Referral created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
  /referrals/code:
    get:
      summary: Get referral by code
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Referral details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
...
  /branches:
    get:
      summary: List branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      summary: Create a branch (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branch"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
...
  /therapists/{id}:
    get:
      summary: Get therapist profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Therapist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Therapist"
  /therapists/{id}/services:
    get:
      summary: Get therapist services
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
  /therapists/services/{service_id}:
    delete:
      summary: Remove service from therapist
      parameters:
        - in: path
          name: service_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service removed
  /therapists/profile:
    patch:
      summary: Update therapist profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                years_experience:
                  type: integer
                certifications:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Therapist"
  /therapists/documents:
    post:
      summary: Upload therapist document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                document_type:
                  type: string
                document_url:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Document uploaded
  /therapists/services:
    post:
      summary: Add service to therapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service_id:
                  type: string
      responses:
        "201":
          description: Service added
  /emergency/trigger:
    post:
      summary: Trigger emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alert_type:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                description:
                  type: string
      responses:
        "201":
          description: Alert triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
  /therapists/{id}/documents:
    get:
      summary: Get therapist documents (Admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      type: object
                      properties:
                        document_id:
                          type: string
                        document_type:
                          type: string
                        status:
                          type: string
  /therapists/documents/{document_id}/verify:
    post:
      summary: Verify therapist document (Admin only)
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [verified, rejected]
                notes:
                  type: string
      responses:
        "200":
          description: Document verified
  /admin/actions:
    get:
      summary: Get admin actions
      responses:
        "200":
          description: List of actions
          content:
            application/json:
              schema:
                type: object
                properties:
                  actions:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminAction"
    post:
      summary: Log admin action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAction"
      responses:
        "201":
          description: Action logged
...
  /emergency/alert/{id}:
    get:
      summary: Get emergency alert details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
...
  /payments/booking/{booking_id}:
    get:
      summary: Get payment by booking ID
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...